backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
celery
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
google-genai
httpx
